+++ daemon-user/src/error.rs
@@ -6,6 +6,7 @@ use rog_anime::error::AnimeError;
 pub enum Error {
     Io(std::io::Error),
     ConfigLoadFail,
+    ConfigLockFail,
     XdgVars,
     Anime(AnimeError),
 }
@@ -16,6 +17,7 @@ impl fmt::Display for Error {
         match self {
             Error::Io(err) => write!(f, "Failed to open: {}", err),
             Error::ConfigLoadFail => write!(f, "Failed to load user config"),
+            Error::ConfigLockFail => write!(f, "Failed to lock user config"),
             Error::XdgVars => write!(f, "XDG environment vars appear unset"),
             Error::Anime(err) => write!(f, "Anime error: {}", err),
         }
@@ -33,5 +35,11 @@ impl From<std::io::Error> for Error {
 impl From<AnimeError> for Error {
     fn from(err: AnimeError) -> Self {
         Error::Anime(err)
+    }
+}
+
+impl From<Error> for zbus::fdo::Error {
+    fn from(err: Error) -> Self {
+        zbus::fdo::Error::Failed(format!("Anime zbus error: {}", err))
     }
 }