+++ daemon-user/src/lib.rs
@@ -1,3 +1,9 @@
 pub mod user_config;
 
 pub mod error;
+
+pub mod ctrl_anime;
+
+pub mod zbus_anime;
+
+pub static DBUS_NAME: &str = "org.asuslinux.Daemon";