+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Revert zbus to 1.9.1
 - Use enum to show power states, and catch missing pci path for nvidia.
 - Partial user-daemon for anime/per-key done, asusd-user. Includes asusd-user systemd unit.
+- user-daemon provides dbus emthods to insert anime actions, remove from index, set leds on/off
 
 # [3.3.0] - 2021-04-3
 ### Changed